requirements.txt
waveshare-epaper
watchdog
pillow
